fix: use a persistent client based on opensource implementation
